Move block building to sequencer
