+++ op-node/rollup/engine/engine_update.go
@@ -6,64 +6,8 @@ import (
 	"fmt"
 
 	"github.com/ethereum-optimism/optimism/op-service/eth"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/rpc"
 )
-
-// isDepositTx checks an opaqueTx to determine if it is a Deposit Transaction
-// It has to return an error in the case the transaction is empty
-func isDepositTx(opaqueTx eth.Data) (bool, error) {
-	if len(opaqueTx) == 0 {
-		return false, errors.New("empty transaction")
-	}
-	return opaqueTx[0] == types.DepositTxType, nil
-}
-
-// lastDeposit finds the index of last deposit at the start of the transactions.
-// It walks the transactions from the start until it finds a non-deposit tx.
-// An error is returned if any looked at transaction cannot be decoded
-func lastDeposit(txns []eth.Data) (int, error) {
-	var lastDeposit int
-	for i, tx := range txns {
-		deposit, err := isDepositTx(tx)
-		if err != nil {
-			return 0, fmt.Errorf("invalid transaction at idx %d", i)
-		}
-		if deposit {
-			lastDeposit = i
-		} else {
-			break
-		}
-	}
-	return lastDeposit, nil
-}
-
-func sanityCheckPayload(payload *eth.ExecutionPayload) error {
-	// Sanity check payload before inserting it
-	if len(payload.Transactions) == 0 {
-		return errors.New("no transactions in returned payload")
-	}
-	if payload.Transactions[0][0] != types.DepositTxType {
-		return fmt.Errorf("first transaction was not deposit tx. Got %v", payload.Transactions[0][0])
-	}
-	// Ensure that the deposits are first
-	lastDeposit, err := lastDeposit(payload.Transactions)
-	if err != nil {
-		return fmt.Errorf("failed to find last deposit: %w", err)
-	}
-	// Ensure no deposits after last deposit
-	for i := lastDeposit + 1; i < len(payload.Transactions); i++ {
-		tx := payload.Transactions[i]
-		deposit, err := isDepositTx(tx)
-		if err != nil {
-			return fmt.Errorf("failed to decode transaction idx %d: %w", i, err)
-		}
-		if deposit {
-			return fmt.Errorf("deposit tx (%d) after other tx in l2 block with prev deposit at idx %d", i, lastDeposit)
-		}
-	}
-	return nil
-}
 
 var ErrEngineSyncing = errors.New("engine is syncing")
 
@@ -80,9 +24,9 @@ const (
 	BlockInsertPayloadErr
 )
 
-// startPayload starts an execution payload building process in the provided Engine, with the given attributes.
+// StartPayload starts an execution payload building process in the provided Engine, with the given attributes.
 // The severity of the error is distinguished to determine whether the same payload attributes may be re-attempted later.
-func startPayload(ctx context.Context, eng ExecEngine, fc eth.ForkchoiceState, attrs *eth.PayloadAttributes) (id eth.PayloadID, errType BlockInsertionErrType, err error) {
+func StartPayload(ctx context.Context, eng ExecEngine, fc eth.ForkchoiceState, attrs *eth.PayloadAttributes) (id eth.PayloadID, errType BlockInsertionErrType, err error) {
 	fcRes, err := eng.ForkchoiceUpdate(ctx, &fc, attrs)
 	if err != nil {
 		var rpcErr rpc.Error
