+++ op-node/rollup/sequencing/errors.go
@@ -1,0 +1,8 @@
+package sequencing
+
+import "errors"
+
+var (
+	ErrSequencerAlreadyStarted = errors.New("sequencer already running")
+	ErrSequencerAlreadyStopped = errors.New("sequencer not running")
+)
